+++ src/main/java/com/example/Catalog/Managment/Repository/OrdersRepository.java
@@ -3,5 +3,5 @@ package com.example.Catalog.Managment.Repository;
 import com.example.Catalog.Managment.Entity.Orders;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrdersRepository extends JpaRepository<Orders, Integer> {
+public interface OrdersRepository extends JpaRepository<Orders, Long> {
 }
